challenge/requirements.txt
pandas
redmail
Jinja2
gspread
oauth2client
python-dotenv